Remove ndk_platform backend. Use the ndk backend.
The ndk_platform backend will soon be deprecated because the ndk backend
can serve the same purpose. This is to eliminate the confusion about
having two variants (ndk and ndk_platform) for the same ndk backend.

The 32-bit/64-bit prebuilts libsmomo.qti.so are patched by directly
editing the string in the file.

Bug: 161456198
Test: m
Change-Id: I03bb468a1d0e0ad90fce39674566386c22e1d9d7