UwbService: Add a module for persisting settings
Add a module for persisting UWB settings to disk. This module will be used for
persisting  simple key/value pairs. It uses a PersistentBundle
underneath for storing the params & for serializing/deserializing the
data to a AtomicFile under the uwb apex data dir.

Since we are maintaing the AOSP -> Vendor stack bridge for now
(pending on b/196225233), we will send the persisted
state to the vendor stack after every reboot.

Bug: 197963882
Test: Toggle UWB state from settings and ensure that the state is
persisted across reboots.
Test: atest com.android.server.uwb

Change-Id: I96bc863a6eb475e4a8ca5aaa0cbe7a528bb5aa79