uwb: Use GET_APP_CONFIG for CCCRangingStarted struct
This is needed to be passed back for CCC sessions via onRangingOpened.
For Fira sessions, we currently return back the FiraOpenSessionParams
that was used for starting the ranging sessions (need to fix this to
return a more appropriate FiraRangingStarted struct in the future).

Note: Very limited manual testing done for the flow with FIRA session
and local modifications. Verified that getAppConfigurations call and generic
traversal of TLV's works. Need to verify the exact set of TLV's for
CccRangingStarted params are parsed correctly or not.

Bug: 208678993
Test: atest CtsUwbTestCases ServiceUwbTests
Change-Id: Ia956ee50c48766a00f2afee108fa94b4b0e3c72e