Dexopt service-uwb.
`service-uwb` is a jar that system_server loads dynamically. Such jars
can be dexopted now. This CL adds `service-uwb` to a
`systemserverclasspath_fragment`, which does not actually add the jar to
the classpath, but tells the build system to dexopt it.

NOTE: The corresponding `exported_systemserverclasspath_fragments` needs
to be added to the `prebuilt_apex` once a new prebuilt is dropped.

Bug: 203198541
Test: m com.android.uwb
Change-Id: I25c429ba47b4bc5b064345ff3488ae5b71e930f7