uwb(uci-jni-rust): Implement vendor uci cmd/rsp/ntf
This is a non-blocking UCI command/response at the API layer. But, it
needs to be synchronous at the UCI layer (since the spec mandates 1
outstanding command to the UWB subsystem). So, use a helper thread to
handle these vendor commands at the java layer without blocking main
java stack.

Bug: 210933436
Test: Compiles
Change-Id: I59e36f88a854b70b8bbc3ff4e07c7bfb01eb2298