uwb(service): Rework reason codes for callbacks
There are 3 sources that need to be converted to API callback reason
codes:
1. UCI status codes
2. UCI reason codes (for session status notification handling)
3. Direct API calls (stopRanging, UWB toggle off).

Handle all of these 3 sources correctly.

Bug: 225068315
Test: atest ServiceUwbTests
Test: Manual tests using field test app:
a)  When stopRanging is called, LOCAL_API reason code is seen.
b)  When UWB is toggled off, SYSTEM_POLICY reason code is seen.
c)  When stopRanging is called on peer controller, REMOTE_REQUEST reason code is seen.

Change-Id: I141984af52ede323376358fbe62b8d9566f4f330