uwb(service): Start timer to track continous errors
Previous approach of using the ranging result callbacks does not work if
the device is the responder and the initiator disappears. The firmware
will stop sending ranging result callbacks in that case.

Also, declare error for ranging result only if all the ranging measurements
(for one to many sessions) are in error.

Bug: 222379826
Test: Manual tests
Test: atest ServiceUwbTests
Change-Id: I869882582cbdb39d70107a54a6ad794decd5f0c9