uwb(support_lib): Add a new builder constructor for ranging open params
Add a new constructor that uses an existing object to create a builder.
This is more safer than existing mechanism because folks
might add new fields to the structure and not see the need for updating
the copy happening in UwbSessionManager. This will be used to update the
cached params for a session whenever it is modified by the app.

1. Switch existing CCC ran interval update logic on session
start to use this new constructor.
2. Update the FIRA session params cache when session is reconfigured.

This will be used by the reconfigure logic to be added for fg state
change handling.

Bug: 235488235
Test: atest ServiceUwbTests
Change-Id: Idad7d4b483fe812b2a44185a7c852a9c4f74144b