uwb(service): Allow reconfiguration of sessions only if hw supports it
Use the capabilities exposed by the hardware to determine if we should
turn off/on notification when the 3p app goes into background.

Also, changed the default for range data notification control to none
(since we have never tested this feature before on any devices even
though it was mandatory in FiRa 1.0 spoc. We want OEMs to opt into this
behavior).

Bug: 235488235
Test: atest ServiceUwbTests
Test: Manually tested the notification setting change when 3p app is moved
to bg/fg using HelloUwb

Change-Id: Iad4ff134dbef6e5adf3b19a97ae092d3b47f302a