Add Android adaptation for uwb core - jni
This is work in progress for Android adaptation for UWB core service. This change adds a jni rust library and a placeholder java UWB service library currently used for local testing. This change contains partial implementation for native rust JNI functions. Callback mechanism isn't implemented as a part of this change.

Test: m libuwb_adaptation_jni; mma; local tests on a device
Bug: 244769936
Change-Id: I79ac2a5eb5665b970821873e184b409205dcb7ac