uwb(framework): Make SessionHandle unique
Curently we use |SessionHandle| to lookup sessionID in the UWB service.
However, |SessionHandle| is not unique. Every instance of
"UwbManager" will have it's own incrementing |SessionHandle|'s for
sessions which causes conflicts in the UWB service session management.

Multiple scenarios that need to be handled for session management across
various apps. Making |SessionHandle| unique is critical for handling all
of the below cases:

1. Different apps opening uwb sessions
   - will have different |attributionSource| & |pid| fields.
2. Same app opening multiple uwb session using same/different "UwbManager"
   instances
   - |Id| field is now static, so different instances of "UwbManager"
     within the same process will now use incrementing |Id|.
3. Different processes with same shared "uid"/"package-name" opening uwb
   sessions
   - will have different |pid| fields.

Bug: 241298182
Bug: 241092129
Test: Manual verification via shell commands.
Test: atest FrameworkUwbTests ServiceUwbTests CtsUwbTestCases
Change-Id: Ib9234f402cf1669b47c083e735e7d1c9f6e4acb0