uwb(service): Wait on country code for stack init.
On the first UWB stack enable, do a non-blocking wait (with a timeout)
for the UWB country code to be known, before returning the UWB stack
enabled notification. On future stack enables, there will not be a need
for the wait as the last known country code will be used.

Bug: 240379493
Test: Manual testing with adb logs
Test: atest ServiceUwbTests
Change-Id: I17b1f0a48e69e8a9d2dbae25ff81588cb90035c3