Store received data packets in their UwbSession
The ReceivedDataInfo represents application payload data received from a
remote device via UCI DATA_MESSAGE_RCV packet(s) (in an active UWBSession).
Hence this should be considered to be a property of the UWBSession, and so
moving it inside the UwbSession class.

The packet(s) are stored in a nested map:
- The outer map's key is the remote device's MacAddress.
- The inner map is a SortedMap, keyed by the packet's sequence number.
- With this representation, we can easily check (and ignore) any duplicates
  while storing them, as well as retrieve the packets in sequence number order,
  for passing them on to the upper layer.

Test: atest ServiceUwbTests FrameworkUwbTests
Bug: 246678053
Change-Id: Idbe7aaa9028c86a5159cb5683203afd922bf6dbf