Implement new UCI CMD to fetch max data size
This CL implments a new UCI CMD defined in CR490
(SESSION_QUERY_DATA_SIZE_IN_RANGING_CMD). This is used to fetch the max
application data payload size that can be sent by the UWBS, in one
ranging round.

This is based on the CL upstreamed by NXP (aosp/2387985).

Bug: 251477752
API-Coverage-Bug: 267698315
Test: atest ServiceUwbTests FrameworkUwbTests CtsUwbTestCases
Change-Id: I3922fde56e9d36709b1bc8a689b34a4d73ed5f94