Do inline AdapterState notify on UWB Disable
Currently, it's possible for AdapterState (callback) notifications for UWB Disable to not be sent out. This happens when a Device Status Notification with DEVICE_STATE_ERROR is received - the handleDisable() called as a part of TASK_RESTART message processing should send an AdapterStateCallback, but it gets over-written by the onCountryCode() callback called within handleEnable()).

The fix is to inline the AdapterState callbacks from handleDisable(),
ie, the callback will happen within that method, rather than enqueuing a
separate TASK_NOTIFY_ADAPTER_STATE message for it.

Enhanced the Device State Error unit test, to include the synchronous
onCountryCode() callback being invoked. Verified that the test was
failing without the fix in this CL.

Change-Id: Idf18887f1309315342b7d64aee97533b770de114
Bug: 270419624
Test: atest ServiceUwbTests