Back Azimuth Primer
Created the back azimuth primer, which detects and correctly reports
when the azimuth has rotated behind the user.

Refactored time handling to use ms since boot instead of Instant, to
make mocking during tests easier and prevent anomalies when the system
clock changes.

Found and fixed a bug in the MedAvgRotationFilter sorting routine that
fails when all values are very close together.

Test: atest CtsUwbTestCases FrameworkUwbTests ServiceUwbTests
Bug: b/275327978 b/272745134
Change-Id: I7d9cc33272b4c413e8f551fdb9426345eeeeb69a