Notify with ReasonCode SYSTEM_REGULATION when countryCode is not valid.
When UWB is enabled and a valid country code is not set, the UWB stack
currently sends an AdapterState notification with State=Disabled and
Reason=SYSTEM_POLICY.

This CL changes the reason to SYSTEM_REGULATION, to match other
scenarios like UWB not being enabled due to the country code. Doing this
change after confirming with the OEMs.

Test: atest ServiceUwbTests
Bug: 267554906
Change-Id: I203175d8628e12820377188627bd23cce20c4d9e