uwb(service): Retry after initializaton failure
Added a listener in UwbServiceCore to avoid creating circular dependency
with UwbServiceImpl. (UwbServiceImpl stores all toggle states, so we need
to check the current toggle states when retrying to avoid cases where
UWB is turned off by the user while waiting for the retry to kick-in).

Bug: 286450197
Test: atest ServiceUwbTests
Test: Manual tests by modifying Qorvo HAL to hang on first
initialization attempt. Verified that the retry was able to recover the
UWB state.

Change-Id: I751b779f4beec1e43dea2b4ad266e033dabff069