Check given params for pause/resume APIs.
Check the given SuspendRangingRounds value in the "params" bundle, for
both the pause() and resume() APIs.

This is required as for both the APIs the given values are used to set
the AppConfig parameter SUSPEND_RANGING_ROUNDS, and so with this check
we can avoid inadvertently triggering the wroing action by one of these APIs.

Bug: 293139823
Test: atest ServiceUwbTests
Change-Id: Ifb6a7c75d0bb11a18afa2b4c9e89bb5fa4b1a168