Notify ranging session closure after removing data.
This CL changes the order between notifying the application about a
ranging session closure and deleting it's data from the UWB
SessionManager.

Currently the UWB Session Manager first notifies the application about
a ranging session closure, and then proceeds to remove the state
associated with it. This can lead to race conditions, as the application
may attempt to create a new ranging session (with same sessionID),
before the SessionManager is able to clear out the state.

Bug: 291291261
Test: atest ServiceUwbTests FrameworkUwbTests CtsUwbTestCases CtsUwbMultiDeviceTestCase_FiraRangingTests
Change-Id: Iba6446eacf03f13f2e373c59323e29f1a2663b48