Send GET_DEVICE_INFO RSP to UWB Service layer
Send the GET_DEVICE_INFO RSP to UWB Service. In a subsequence CL, it
will be stored and used for checking the UWBS UCI version, that should
change the way UWB_INITIATION_TIME is configured.

This is a roll-forward of the reverted CL aosp/2697893, with the merge conflict fixed.

Bug: 276825370
Test: atest ServiceUwbTests
Test: manual testing (verified UWB is enabled and a proper
DeviceInfoResponse is received by UwbServiceCore).

Change-Id: I6d4b939177824c65ed8a1c3d8d682b123b33aaf5