[uwb] Add a flag to force updating cached specification params
If country code is changed and open ranging coincides, open ranging
recives a non-null params which will be null if there is a change in
country code.

With a flag in place, we never set cached params to null but will
force getting a new spec if the flag in true.

Bug: 295109521
Test: atest ServiceUwbTests
Change-Id: I269c2abd95a6e78e920a29ecb5c84f74148a85f5