Set initiation_time_ms to 0ms in tests.
We expect the UWB_INITIATION_TIME to be set to 0ms in the real-world use
cases (and it is not modifiable by the 3p API), so setting it to the
default value of 0ms in these tests.

Verified in the CTS test run that UwbInitiationTime is set to 0 for all
the test cases:
"(AppConfigTlv { cfg_id: UwbInitiationTime, v: [0, 0, 0, 0] })".

Bug: 276932800
Test: atest CtsUwbTestCases CtsUwbMultiDeviceTestCase_UwbManagerTests CtsUwbMultiDeviceTestCase_FiraRangingTests
Change-Id: I71166895ad600f7a3f25ed2193ccaf34e633a346