Replace FiraParams.mProtocolVersion with UWBS UCI version.
Currently FiraOpenSessionParams contains a field "mProtocolVersion" that
signifies the UCI version for which App Config parameters need to be
setup. This field is set by the UWB app or UWB ConfigurationManager, which is error-prone behavior.

This CL replaces it's usage with the UCI version returned by UWBS (in CORE_GET_DEVICE_INFO_RSP).

Test: atest ServiceUwbTests
Test: Verified the new flag is getting enabled with the command "adb shell device_config put uwb com.android.uwb.flags.use_uwbs_uci_version true".
Bug: 303104817
Change-Id: I7f18a73e4870191191db5e8bcc9e8cb3c8e130ab