Replace FiraParams.mProtocolVersion with UWBS UCI version.
Currently FiraOpenSessionParams contains a field "mProtocolVersion" that
signifies the UCI version for which App Config parameters need to be
setup. This field is set by the UWB app or UWB ConfigurationManager, which is error-prone behavior.

This CL replaces it's usage with the UCI version returned by UWBS (in CORE_GET_DEVICE_INFO_RSP).

This is cherry-picked from the CL aosp/2772286, with all the code
related to Android feature flags removed (since they are not supported on udc-mainline-prod).

Test: atest ServiceUwbTests
Bug: 303104817
Merged-In: I7f18a73e4870191191db5e8bcc9e8cb3c8e130ab
Change-Id: I0607adade8e0e5c0c53b2060217b613a05e41ce2