Add a flag to enable Absolute UWB Initiation Time for CCC.
The flag controls whether the UWB initiation time in the
CccStartRangingParams is directly configured in the UWBS (a relative
time), or, an absolute UWB initiation time is computed using it, for a
CCC ranging session, on a Fira UCI 2.0+ device.

The flag is declared with default of "false", which maps to the former
behavior (of being used as a relative time).

Bug: 307151218
Test: atest ServiceUwbTests
(cherry picked from https://android-review.googlesource.com/q/commit:a9e6d643cbcbbbc68399e9a615ac9f621ca52155)
Merged-In: If21296e8f53981abfd2ebf1cecd10e005d4d5bf9

Change-Id: If21296e8f53981abfd2ebf1cecd10e005d4d5bf9