Configure absolute UWB_INITIATION_TIME for CCC session.
The SET_APP_CONFIG CMD parameter UWB_INITIATION_TIME(0x2B) is defined as
an absolute timestamp for Fira 2.0+ devices. Currently however it is
being configured as the given (relative) time, at the time of a CCC
ranging session initialization. This CL fixes the behavior (like was
earlier done for Fira ranging sessions in aosp/2746407).

This change is protected behind a DeviceConfigFacade flag, that controls
whether a relative or absolute UWB_INITIATION_TIME should be configured,
for CCC ranging sessions.

Bug: 291851851
Test: atest ServiceUwbTests
Change-Id: I4af9f75ab0f93c068df571aa65b60b36f0736864