[Uwb] Compute the abs initiation time at start even if null params
It is possible/reasonable to set a relative initiation time in the
`{Ccc,Aliro}OpenRangingParams`. If
`ccc_absolute_uwb_initiation_time_enabled` is set then it's expected
that this relative time will be converted to an absolute time and sent
via `SET_APP_CONFIG` before the session init time. This all works
fine. See r.android.com/2746407 for reference.

After session init time, the computed absolute initiation time will be
set back to 0. See the `resetAbsoluteInitiationTime()` call in
`handleOpenRanging()`, again referencing r.android.com/2746407. It's
then expected that we'll re-compute and re-send an updated absolute
initiation time at session start time.

When we get to start ranging, it's possible that some parameters are
overridden. See `{Ccc,Aliro}StartRangingParams()`. ...but it's also
possible that we don't want to override any parameters. It can be
noted that in `UwbSessionManager` that the `startRanging()` method
specifies that the params are `@Nullable`. When no specific params
overrides are passed to `startRanging()`, the code will presist the
values from the `{Ccc,Aliro}OpenRangingParams`.

Due to an oversight in the code, we won't try to re-compute the
absolute initiation time when `startRanging()` is called with null
`{Ccc,Aliro}StartRangingParams()`. We'll copy over the relative
initiation time from the `{Ccc,Aliro}OpenRangingParams` but the
absolute initiation time will be left as 0. This means that when we
later call `SET_APP_CONFIG` after the session has started that we'll
do so with the relative initiation time set but with the absolute
initiation time left at 0. Due to the way that the `CccEncoder` works
currently (which arguably is a bug itself) this relative initiation
time is sent via `SET_APP_CONFIG`.

Fix the oversight and properly re-compute the absolute initiation time
when `startRanging()` is called with null
`{Ccc,Aliro}StartRangingParams()`. While touching the code, also make
sure to set `mNeedsAppConfigUpdate` in the generic FiRa code if we
need to compute the absolute time. This isn't fixing any known bugs
but matches the updated logic in the Ccc and Aliro cases.

Said another way, start with these pre-conditions:
* `ccc_absolute_uwb_initiation_time_enabled` is set.
* `{Ccc,Aliro}OpenRangingParams.initiationTimeMs` is set to non-zero.
* `{Ccc,Aliro}OpenRangingParams.absoluteInitiationTimeUs` is zero.
* `startRanging()` is called with null parameters, which currently
  happens with the digital car key code when ranging is opened.

Before ${SUBJECT} patch:
* We'll correctly send the absolute initiation time (via
  `SET_APP_CONFIG`) at session init time.
* We'll incorrectly send the relative initiation time (via
  `SET_APP_CONFIG`) instead of sending an updated absolute initiation
  time at session opened time.

After ${SUBJECT} patch:
* (same) We'll correctly send the absolute initiation time (via
  `SET_APP_CONFIG`) at session init time.
* We'll correctly send an updated absolute initiation time at session
  opened time.

Bug: 425468690
Flag: EXEMPT bug fix
Test: Setup pre-conditions and see abs init time is right in logs
Change-Id: Icf3909fac4b7bb5a5a3239a74e51be19af14ac1c