[Uwb] Make minUwbInitiationTimeMs relevant (redux)
The `{Ccc,Aliro}SpecificationParams` (which effectively comes from the
HAL and/or firmware) have the concept of a `minUwbInitiationTimeMs`.
This was based on an Android-specific capability added in
r.android.com/2523137 ("Add min uwb initiation time as new vendor
capability") and then added to the Ccc params in r.android.com/2527924
("Add min start delay to CCC spec params").

Nothing seems to use this time.

Start using the `minUwbInitiationTimeMs` if/when we convert from a
relative initiation time to an absolute one. This will be used to
enforce that the releative initiation time we use is never less than
the minimum that comes from thue HAL/firmware. The only time we
convert from a relative initiation time to an absolute one appears to
be if the `uwb` `ccc_absolute_uwb_initiation_time_enabled` property is
set, so this patch shouldn't do anything on devices that don't set
that.

NOTES:
* This interpretion of the delays means that, with the way the Android
  frameworks currently work, both the `initiationTimeMs` that's part
  of `{Ccc,Aliro}CccOpenRangingParams` and the `MIN_INITIATION_TIME`
  that the from the HAL/firmware both represent the *total needed
  delay* from the time the session is started until we should transmit
  the first packet over UWB. This total delay needs to account for the
  startup time of the UWB chip, the startup time of the vehicle, the
  time to transmit a Bluetooth packet to the vehicle, and any
  margins/overheads.
* If an application is using an absolute UWBS initiation time (which
  would be awkward) the minimum won't be enforced. We _could_ do all
  the math and enforce this still if needed, but it would be more code
  and isn't part of this patch.

The reland of the change properly checks the type of the params to
figure out which spec to use rather than just using the first non-NULL
Ccc/Aliro spec.

Bug: 425468690
Bug: 438521042
Flag: EXEMPT bug fix
Test: In logs, see that this overrides a smaller value
Change-Id: I877d3950a2da30d00a8174e9dd82896011e13687