[Uwb] Avoid a few extra constructions of {Ccc,Aliro}OpenRangingParams
In openRanging(), we need to do the following:
* Parse the parameters.
* Look at the result of parsing and see if hopping mode needs to be
  setup. If so, set the hopping mode key.

The design of {Ccc,Aliro}OpenRangingParams is for the class to be
immutable. To change something, you create a builder from an object,
set things in the builder, and build a new copy of the object with the
changes. This means that if we need to set the hopping mode key we
need to contruct the {Ccc,Aliro}OpenRangingParams object twice: once
during parsing so we can check if hopping mode needs to be setup and
once after we've set the hopping mode key. If we don't need to set the
hopping mode key we should only construct the object once.

Right now, we are constructing the object three times in all
cases. Fix that.

The first extra construction is when we create the builder. There's no
need to parse a second time and give that to the Builder
constructor. The builder makes a copy of all parameters in the old
params.

The second extra construction is when we don't need to set the hopping
mode key. This should be in the conditional.

This change is expected to be a noop, but it should make the code
slightly more efficient and (more slightly easier to understand).

Bug: 427218733
Flag: EXEMPT minor cleanup
Test: Time Sync testing
Change-Id: I167f5127f872b8ae7a26dea03c81a9dd17002159