[Uwb] Don't override initiationTimes from start parms if they are 0
The initiation time values (`absoluteInitiationTimeUs` and
`initiationTimeMs`) are present in both `{Ccc,Aliro}OpenRangingParams`
and `{Ccc,Aliro}StartRangingParams`. This leads to the question of
which the value needs to be specified in.

In general it's probably a good idea to specify it in _at least_
`{Ccc,Aliro}OpenRangingParams` due to issues talked about in
r.android.com/2746407 (namely that some UWBS check for a valid value
at SESSION_INIT time), but that's not strictly required by the code,
which works fine with the value specified _only_ in
`{Ccc,Aliro}StartRangingParams` (assuming the UWBS isn't upset by
this).

It's unclear if it's OK to specify the value _only_ in
`{Ccc,Aliro}OpenRangingParams` but _not_ in
`{Ccc,Aliro}StartRangingParams`. Given that it's possible to start a
session _without_ passing `{Ccc,Aliro}StartRangingParams`, this is
it's probably expected to work. It would be unfortunate if needing to
update the RAN multiplier (which meant you couldn't use null params)
suddenly meant that you needed to manually copy the initiation times
from the `{Ccc,Aliro}OpenRangingParams` to the
`{Ccc,Aliro}StartRangingParams`. Thus, assume that it's OK to _only_
specify the params in `{Ccc,Aliro}OpenRangingParams`.

Given the above conclusion, update the code so that if the initiation
time values in `{Ccc,Aliro}StartRangingParams` are zero that we don't
clobber the old values in `{Ccc,Aliro}OpenRangingParams`.

Said another way, start with these pre-conditions:
* One of `{Ccc,Aliro}OpenRangingParams.initiationTimeMs` or
  `{Ccc,Aliro}OpenRangingParams.absoluteInitiationTimeUs` is non-zero.
* `startRanging()` is called with non-null parameters, but
  `initiationTimeMs` and `absoluteInitiationTimeUs` are both zero in the
  parameters passed to `startRanging()`

Before ${SUBJECT} patch:
* We'll correctly send the initiation time (via `SET_APP_CONFIG`) at
  session init time.
* We'll incorrectly clobber the initiation time at session opened time.

After ${SUBJECT} patch:
* (same) We'll correctly send the initiation time (via `SET_APP_CONFIG`)
  at session init time.
* We'll correctly keep the initiation time at session opened time.

Bug: 425468690
Flag: EXEMPT bug fix
Test: Setup pre-conditions and see abs init time is right in logs
Change-Id: Ia87bf4338e40c560a2f91af07431b60822d11fd6