[Uwb] Copy initiation times in {Ccc,Aliro}OpenRangingParams Builder ctor
The contructor from an existing params for the `Builder` in
`{Ccc,Aliro}OpenRangingParams` doesn't copy the `mInitiationTimeMs`
nor the `mAbsoluteInitiationTimeUs`. This looks to be a straight
oversight from when initiation time was first added in
r.android.com/2540991, as there is no reason you wouldn't want to
preserve these values when using a builder to created a modified
version of these parameters.

While fixing the oversight, also move location of setting of
`mStsIndex` to make it so that diffs between the
constroctor-from-old-Builder and constructor-from-old-Params methods
match each other.

Bug: 425468690
Flag: EXEMPT bug fix
Test: Time Sync testing
Change-Id: I134754860e7842910de1f6b90560f6e46c5a683b