Use location = UNKNOWN & sight = NLOS as GRAPI OOB CS defaults
We're currently not exposing the ability for the caller to select these
params when in OOB mode, but we do for raw mode. For OOB mode it's ok to
make the defaults UNKNOWN and NLOS as opposed to the current defaults of
UNKNOWN & UNKNOWN, since the UNKNOWN,NLOS pair will use the cs alg
that's applicable for the most common case

Later on if needed we might want to expose these params to the user, but
no need currently.

Bug: 389138805
Flag: EXEMPT minor change
Test: defaults change
Change-Id: I96a39e01ff14b87e7246a995e6eae24f6248ab15