uwb(metrics): Attribute UWB metrics to non-privileged apps when present.
This change introduces `getAttributionSourceForMetrics` to select the appropriate `AttributionSource` for logging. If a UWB session involves a non-privileged application, the metrics will be attributed to one of those non-privileged apps. Otherwise, the primary session attribution source is used. This ensures better privacy and accountability in multi-app UWB sessions.

Bug: 406597556
Test: Compiles
Flag: EXEMPT bug fix
Change-Id: I9e8ee89c8307e34e9ab2168533a61d60e4cc877d