[Ranging] Determine initiator and responder role on UWB sessions
Introduces the concept of device type to UWB ranging sessions to
dynamically determine the initiator and responder roles.
Previously, the roles were statically assigned. With this change,
the roles are determined based on a power ranking of the device types.

This change modifies the ConfigSelector interface to include the
deviceType when adding peer capabilities. The UwbConfigSelector
now uses this information to select the appropriate role for
each peer in the ranging session.

Bug: 460157518
Test: manual with RangingTestApp
Flag: EXEMPT BUGFIX
Change-Id: I52d917ded8da2d5f90528423a19430b290475e0b